add c3x3_16 operator
it's a 3x3 convolution op, it will expand the filter number to 15x.
each expand filter is a rotate or a mirror of the origin filter.

why this op?

after analysis the invariance of the SIFT algorithm, and compare it
with cnn net, found this op may help the cnn net to converge faster.

tested by c3x316test.py.